Close loophole with possible nested history objects.
I don't want to deal with this. The end user doesn't want to deal with this. And it's headache inducing to even have to think about this or manage this in any class that's imposing a wrapping of the converse history framework. So we're just going to address this at the root. Nested history objects is not likely to be a feature anyone wants, so preventing it here and now saves everyone a lot of confusion and frustration.